Refactor offline recognizer
